refine topic modelling
