+++ src/config.yaml
@@ -2,15 +2,3 @@ raw_data_path: "data/raw/20230717_consultation_ingest.csv" #str
 additional_stopwords: #list of words to filter; must be type str
   - 'he'
 lemmatize: True #bool; select False to use Stemmer
-feature_count: #dict
-  ngram_range: !!python/tuple [1,2] #tuple range of defaults to unigram (1,1)
-  min_df: 2 #float (proportion) or int (count)
-  max_df: 0.9 #float (proportion) or int (count)
-  max_features: null #null converts to None, or int value
-  lowercase: True #whether to convert all words to lowercase
-lda: #dict
-  n_topics: 3 #int greater than 0
-  n_top_words: 10 #int
-  max_iter: 25 #int
-  title: "Topic Summary" #str
-  topic_labels: null # also takes a list of strings (see additional stopwords ^)
