--- src/question_config.yaml
@@ -1,11 +1,0 @@
-questions_to_interpret:
-  - 12
- # - 14
- # - 16
- # - 19
- # - 21
- # - 24
- # - 28
- # - 30
- # - 31
- # - 32
